Ajouter au tableau de bord la possibilité de gérer l'état d’une commande

+++ database/migrations/2025_04_21_220235_add_payment_method_to_orders_table.php
@@ -26,8 +26,11 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::table('orders', function (Blueprint $table) {
-            $table->dropColumn('payment');
-        });
+        if (Schema::hasColumn('orders', 'payment') ){
+            Schema::table('orders', function (Blueprint $table) {
+                $table->dropColumn('payment');
+            });
+        };
     }
 };
+
